OCR/pyproject.toml
[tool.poetry]
package-mode = false
name = "ocr"
version = "0.1.0"
description = "Digitize scanned PDFs"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
python-dotenv = "^1.0.1"
transformers = {extras = ["torch"], version = "^4.39.3"}
pillow = "^10.3.0"
pypdf = "^3.0.1"
pymupdf = "^1.24.2"
pytest = "^8.1.1"
pdf2image = "^1.17.0"
pytest-mock = "^3.14.0"
pytesseract = "^0.3.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "ocr.main:main"

[tool.ruff]
line-length = 118
target-version = "py310"
